+++ packages/sdk/CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Unreleased
 
 - Change note commitment to use encoded `asset` and `ID`
-- Add `fromEncoded` and `toEncoded` methods to `NoteTrait`
+- Add and `encode` and `decode` methods to `NoteTrait`
 - Change `NoteInput` to `EncodedNote`
 - Remove `toJSON` and all `fromJSON` methods in favor of custom `bigint-json-serialization`
 - Add query function for `SubtreeUpdate` events